mf-karr-backend/scripts/indexUpdate/requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0 
flask_cors>=5.0.0
requests==2.31.0
psycopg2-binary==2.9.10
yfinance>=0.2.55